replaced _specify_platform_settings_type for consistency
